+++ packages/theme/pigeonposse/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dovenv/theme-pigeonposse
 
+## 1.5.32
+
+### Patch Changes
+
+- up deps, fix issues in theme-pigeonposse
+
+- Updated dependencies []:
+  - @dovenv/stylelint-config@1.5.32
+  - @dovenv/eslint-config@1.5.32
+  - @dovenv/theme-banda@1.5.32
+  - @dovenv/core@1.5.32
+
 ## 1.5.31
 
 ### Patch Changes